requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
requests==2.31.0
email-validator==2.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
PyJWT==2.10.1
cryptography==41.0.7
disposable-email-domains==0.0.126
supabase==2.17.0
feedparser==6.0.11
podcast-api==1.1.6
APScheduler==3.10.4
pydantic>=2.11.7,<3.0.0
itsdangerous==2.1.2
starlette==0.27.0
boto3>=1.26.0
botocore>=1.29.0
Pillow>=9.0.0
ffmpeg-python>=0.2.0
weasyprint>=60.0
markdown2>=2.4.0
ics>=0.7.2
google-generativeai>=0.3.0
